splitting up test/krl.ts
